+++ app/kuma-cp/cmd/cmd_k8s_test.go
@@ -74,6 +74,6 @@ guiServer:
   port: 0
 `,
 			admissionServerPort,
-			filepath.Join("..", "..", "kuma-injector", "cmd", "testdata"))
+			filepath.Join("testdata"))
 	}))
 })