+++ pkg/plugins/runtime/k8s/plugin.go
@@ -4,6 +4,8 @@ import (
 	"fmt"
 	"strings"
 
+	"github.com/Kong/kuma/pkg/plugins/runtime/k8s/webhooks/injector"
+
 	"github.com/pkg/errors"
 
 	"github.com/Kong/kuma/pkg/core"
@@ -22,7 +24,7 @@ import (
 
 	kube_schema "k8s.io/apimachinery/pkg/runtime/schema"
 	kube_ctrl "sigs.k8s.io/controller-runtime"
-	kuba_webhook "sigs.k8s.io/controller-runtime/pkg/webhook"
+	kube_webhook "sigs.k8s.io/controller-runtime/pkg/webhook"
 )
 
 var (
@@ -51,6 +53,8 @@ func (p *plugin) Customize(rt core_runtime.Runtime) error {
 		return err
 	}
 
+	addMutators(mgr, rt)
+
 	return addDefaulters(mgr)
 }
 
@@ -66,7 +70,7 @@ func addNamespaceReconciler(mgr kube_ctrl.Manager, rt core_runtime.Runtime) erro
 		Client:              mgr.GetClient(),
 		Log:                 core.Log.WithName("controllers").WithName("Namespace"),
 		SystemNamespace:     rt.Config().Store.Kubernetes.SystemNamespace,
-		CNIEnabled:          rt.Config().Runtime.Kubernetes.CNIEnabled,
+		CNIEnabled:          rt.Config().Runtime.Kubernetes.Injector.CNIEnabled,
 		ResourceManager:     rt.ResourceManager(),
 		DefaultMeshTemplate: rt.Config().Defaults.MeshProto(),
 	}
@@ -125,14 +129,23 @@ func addValidators(mgr kube_ctrl.Manager, rt core_runtime.Runtime) error {
 	mgr.GetWebhookServer().Register(path, composite.WebHook())
 	log.Info("Registering a validation composite webhook", "path", path)
 
-	mgr.GetWebhookServer().Register("/validate-v1-service", &kuba_webhook.Admission{Handler: &k8s_webhooks.ServiceValidator{}})
+	mgr.GetWebhookServer().Register("/validate-v1-service", &kube_webhook.Admission{Handler: &k8s_webhooks.ServiceValidator{}})
 	log.Info("Registering a validation webhook for v1/Service", "path", "/validate-v1-service")
 
 	secretValidator := &k8s_webhooks.SecretValidator{
 		Client: mgr.GetClient(),
 	}
-	mgr.GetWebhookServer().Register("/validate-v1-secret", &kuba_webhook.Admission{Handler: secretValidator})
+	mgr.GetWebhookServer().Register("/validate-v1-secret", &kube_webhook.Admission{Handler: secretValidator})
 	log.Info("Registering a validation webhook for v1/Secret", "path", "/validate-v1-secret")
 
 	return nil
+}
+
+func addMutators(mgr kube_ctrl.Manager, rt core_runtime.Runtime) {
+	kumaInjector := injector.New(
+		rt.Config().Runtime.Kubernetes.Injector,
+		rt.Config().ApiServer.Catalog.ApiServer.Url,
+		mgr.GetClient(),
+	)
+	mgr.GetWebhookServer().Register("/inject-sidecar", k8s_webhooks.PodMutatingWebhook(kumaInjector.InjectKuma))
 }