+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # CHANGELOG
 
 ## master
+* fix: envoy binary not found
+  [#695](https://github.com/Kong/kuma/pull/695)
+  👍contributed by @tharun208
+* feat: merge injector into kuma-cp
+  [#701](https://github.com/Kong/kuma/pull/701)
 * feat: refactor other pars of the Mesh to be consistent with CA
   [#704](https://github.com/Kong/kuma/pull/704)
   ⚠️ warning: breaking change of Mesh model